pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deepcall"
version = "0.1.0"
description = "High-performance real-time communication platform"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "redis>=5.0.1",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    "livekit-api>=0.1.2",
    "arq>=0.25.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.0.post1"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
